Rename barang controller with PSR-12 formatting
